nvapi: Move resource-factory
