+++ src/meson.build
@@ -2,6 +2,7 @@ nvapi_src = files([
   'util/util_string.cpp',
   'util/util_env.cpp',
   'util/util_log.cpp',
+  'nvapi/resource_factory.cpp',
   'nvapi/lfx.cpp',
   'nvapi/nvapi_d3d_instance.cpp',
   'nvapi/nvapi_d3d_low_latency_device.cpp',
@@ -12,7 +13,6 @@ nvapi_src = files([
   'nvapi/nvapi_output.cpp',
   'nvapi/vulkan.cpp',
   'nvapi/nvml.cpp',
-  'resource_factory.cpp',
   'nvapi_globals.cpp',
   'nvapi.cpp',
   'nvapi_d3d.cpp',
