+++ src/nvapi_globals.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "nvapi_private.h"
-#include "resource_factory.h"
+#include "nvapi/resource_factory.h"
 #include "nvapi/nvapi_adapter_registry.h"
 #include "nvapi/nvapi_d3d_instance.h"
