+++ tests/meson.build
@@ -2,6 +2,7 @@ nvapi_src = files([
   '../src/util/util_string.cpp',
   '../src/util/util_env.cpp',
   '../src/util/util_log.cpp',
+  '../src/nvapi/resource_factory.cpp',
   '../src/nvapi/vulkan.cpp',
   '../src/nvapi/nvml.cpp',
   '../src/nvapi/lfx.cpp',
@@ -10,7 +11,6 @@ nvapi_src = files([
   '../src/nvapi/nvapi_output.cpp',
   '../src/nvapi/nvapi_adapter.cpp',
   '../src/nvapi/nvapi_adapter_registry.cpp',
-  '../src/resource_factory.cpp',
   '../src/nvapi/nvapi_d3d11_device.cpp',
   '../src/nvapi/nvapi_d3d12_device.cpp',
   '../src/nvapi_globals.cpp',
